Rename amt -> intelamt:
Per PR request, this add the hardware
vendor's name to the package.
